convert images to JPEG and remove large files
