+++ src/App.css
@@ -7,6 +7,9 @@
 
 /* Universal Styles */
 
+.page-break {
+  height: 100vh;
+}
 h1 {
   font-size: 42px;
 }
